refactor meal form utils, api
