+++ src/features/meal-form/api/createMealCategories.ts
@@ -1,0 +1,33 @@
+import { GraphQLQuery } from '@aws-amplify/api'
+import { API } from 'aws-amplify'
+
+import {
+  CreateMealCategoryInput,
+  CreateMealCategoryMutation,
+  MealCategoryName,
+} from '../../../API'
+import { createMealCategory } from '../../../graphql/mutations'
+
+export async function createMealCategories(mealDateId: string) {
+  try {
+    const mealCategoryNames: MealCategoryName[] =
+      Object.values(MealCategoryName)
+
+    await Promise.all(
+      mealCategoryNames.map(async (name) => {
+        const createMealCategoryInput: CreateMealCategoryInput = {
+          name,
+          mealdateID: mealDateId,
+        }
+
+        await API.graphql<GraphQLQuery<CreateMealCategoryMutation>>({
+          query: createMealCategory,
+          variables: { input: createMealCategoryInput },
+          authMode: 'AMAZON_COGNITO_USER_POOLS',
+        })
+      })
+    )
+  } catch (err) {
+    console.log('Error creating MealCategory:', err)
+  }
+}
